requirements.txt
flask
gunicorn
binance-connector
pyTelegramBotAPI
urllib3==1.21.1
